refactor: Clean up tests
